requirements.txt
openai~=1.54.3
chromadb~=0.5.18
sentence-transformers~=3.3.0
PyYAML~=6.0.2
faiss_cpu
faiss-cpu~=1.9.0
numpy~=2.1.3
h11~=0.14.0
pillow~=11.0.0
pip~=24.2
distro~=1.9.0
protobuf~=5.28.3
filelock~=3.16.1
rsa~=4.9
pyasn1~=0.6.1
grpcio~=1.67.1
rich~=13.9.4
Pygments~=2.18.0
markdown-it-py~=3.0.0
requests~=2.32.3
colorama~=0.4.6
zipp~=3.21.0
sniffio~=1.3.1
importlib_metadata~=8.5.0
pyproject_hooks~=1.2.0
build~=1.2.2.post1
typing_extensions~=4.12.2
torch~=2.5.1
scipy~=1.14.1
setuptools~=75.4.0
click~=8.1.7
httpcore~=1.0.6
idna~=3.10
certifi~=2024.8.30
jiter~=0.7.0
mdurl~=0.1.2
charset-normalizer~=3.4.0
overrides~=7.7.0
regex~=2024.11.6
threadpoolctl~=3.5.0
mpmath~=1.3.0
sympy~=1.13.1
onnxruntime~=1.20.0
tqdm~=4.67.0
packaging~=24.2
urllib3~=2.2.3
fsspec~=2024.10.0
typer~=0.13.0
shellingham~=1.5.4
pyasn1_modules~=0.4.1
cachetools~=5.5.0
MarkupSafe~=3.0.2
joblib~=1.4.2
httpx~=0.27.2
pydantic~=2.9.2
anyio~=4.6.2.post1
pydantic_core~=2.23.4
orjson~=3.10.11
PyPika~=0.48.9
asgiref~=3.8.1
backoff~=2.2.1
fastapi~=0.115.4
starlette~=0.41.2
websockets~=14.0
six~=1.16.0
python-dateutil~=2.9.0.post0
posthog~=3.7.0
monotonic~=1.6
scikit-learn~=1.5.2
uvicorn~=0.32.0
httptools~=0.6.4
watchfiles~=0.24.0
importlib_resources~=6.4.5
tenacity~=9.0.0
bcrypt~=4.2.0
kubernetes~=31.0.0
mmh3~=5.0.1
chroma-hnswlib~=0.7.6
networkx~=3.4.2
oauthlib~=3.2.2
Deprecated~=1.2.14
annotated-types~=0.7.0
Jinja2~=3.1.4
websocket-client~=1.8.0
pyreadline3~=3.5.4
wrapt~=1.16.0
durationpy~=0.9
requests-oauthlib~=2.0.0
tokenizers~=0.20.3
humanfriendly~=10.0
coloredlogs~=15.0.1
flatbuffers~=24.3.25
transformers~=4.46.2
safetensors~=0.4.5
huggingface-hub~=0.26.2
langchain>=0.3,<0.4
langchain-community>=0.3,<0.4